+++ semver-project-gradle-plugin/main/kotlin/com/javiersc/semver/project/gradle/plugin/valuesources/VersionValueSource.kt
@@ -1,12 +1,16 @@
 package com.javiersc.semver.project.gradle.plugin.valuesources
 
 import com.javiersc.gradle.version.GradleVersion
+import com.javiersc.semver.project.gradle.plugin.SemverExtension
 import com.javiersc.semver.project.gradle.plugin.internal.calculatedVersion
 import com.javiersc.semver.project.gradle.plugin.internal.checkCleanProperty
 import com.javiersc.semver.project.gradle.plugin.internal.checkVersionIsHigherOrSame
 import com.javiersc.semver.project.gradle.plugin.internal.commitsMaxCount
 import com.javiersc.semver.project.gradle.plugin.internal.git.GitCache
 import com.javiersc.semver.project.gradle.plugin.internal.git.GitRef
+import com.javiersc.semver.project.gradle.plugin.internal.git.currentBranch
+import com.javiersc.semver.project.gradle.plugin.internal.git.headCommit
+import com.javiersc.semver.project.gradle.plugin.internal.git.lastVersionTagInCurrentBranch
 import com.javiersc.semver.project.gradle.plugin.internal.projectTagPrefix
 import com.javiersc.semver.project.gradle.plugin.internal.scopeProperty
 import com.javiersc.semver.project.gradle.plugin.internal.stageProperty
@@ -15,6 +19,7 @@ import com.javiersc.semver.project.gradle.plugin.semverExtension
 import com.javiersc.semver.project.gradle.plugin.tasks.CreateSemverTagTask
 import com.javiersc.semver.project.gradle.plugin.tasks.PushSemverTagTask
 import java.io.File
+import org.eclipse.jgit.api.Git
 import org.gradle.api.Project
 import org.gradle.api.provider.Property
 import org.gradle.api.provider.Provider
@@ -62,12 +67,25 @@ internal abstract class VersionValueSource : ValueSource<String, VersionValueSou
                     lastVersionCommitInCurrentBranch = lastVersionCommitInCurrentBranch,
                 )
 
-            checkVersionIsHigherOrSame(version, lastSemver)
+            val gradleVersion = GradleVersion(version)
+            val mappedGradleVersion: GradleVersion =
+                versionMapper.get().map(gradleVersion).let(GradleVersion::invoke)
 
-            version
+            val gitData: SemverExtension.GitData = cache().git.buildGitData(tagPrefixProperty)
+            val mappedVersionAndGit: String =
+                versionAndGitMapper.get().map(mappedGradleVersion, gitData)
+
+            checkVersionIsHigherOrSame(
+                version = mappedVersionAndGit,
+                lastVersionInCurrentBranch = lastSemver,
+            )
+
+            mappedVersionAndGit
         }
 
     internal interface Params : ValueSourceParameters {
+        val versionMapper: Property<SemverExtension.VersionMapper>
+        val versionAndGitMapper: Property<SemverExtension.VersionAndGitMapper>
         val gitDir: Property<File>
         val commitsMaxCount: Property<Int>
         val tagPrefixProperty: Property<String>
@@ -83,6 +101,9 @@ internal abstract class VersionValueSource : ValueSource<String, VersionValueSou
         fun register(project: Project): Provider<String> =
             project.providers.of(VersionValueSource::class) { valueSourceSpec ->
                 val gitDir = project.provider { project.semverExtension.gitDir.get().asFile }
+                valueSourceSpec.parameters.versionMapper.set(project.semverExtension.versionMapper)
+                valueSourceSpec.parameters.versionAndGitMapper.set(
+                    project.semverExtension.versionAndGitMapper)
                 valueSourceSpec.parameters.gitDir.set(gitDir)
                 val commitsMaxCount: Int =
                     project.commitsMaxCount.orNull ?: project.semverExtension.commitsMaxCount.get()
@@ -102,3 +123,53 @@ private val Project.isCreatingSemverTag: Boolean
         gradle.startParameter.taskNames.any { taskName: String ->
             taskName == CreateSemverTagTask.NAME || taskName == PushSemverTagTask.NAME
         }
+
+private fun Git.buildGitData(tagPrefix: Property<String>): SemverExtension.GitData {
+    return SemverExtension.GitData(
+        tag =
+            lastVersionTagInCurrentBranch(tagPrefix.get())?.let { tag ->
+                SemverExtension.GitData.Tag(
+                    name = tag.name,
+                    refName = tag.refName,
+                    commit =
+                        SemverExtension.GitData.Commit(
+                            message = tag.commit.message,
+                            fullMessage = tag.commit.fullMessage,
+                            hash = tag.commit.hash,
+                        ),
+                )
+            },
+        commit =
+            SemverExtension.GitData.Commit(
+                message = headCommit.commit.message,
+                fullMessage = headCommit.commit.fullMessage,
+                hash = headCommit.commit.hash,
+            ),
+        branch =
+            SemverExtension.GitData.Branch(
+                name = currentBranch.name,
+                refName = currentBranch.refName,
+                commits =
+                    currentBranch.commits.map { commit ->
+                        SemverExtension.GitData.Commit(
+                            message = commit.message,
+                            fullMessage = commit.fullMessage,
+                            hash = commit.hash,
+                        )
+                    },
+                tags =
+                    currentBranch.tags.map { tag ->
+                        SemverExtension.GitData.Tag(
+                            name = tag.name,
+                            refName = tag.refName,
+                            commit =
+                                SemverExtension.GitData.Commit(
+                                    message = tag.commit.message,
+                                    fullMessage = tag.commit.fullMessage,
+                                    hash = tag.commit.hash,
+                                ),
+                        )
+                    },
+            ),
+    )
+}