+++ semver-project-gradle-plugin/testFixtures/kotlin/com/javiersc/semver/project/gradle/plugin/Sandbox.kt
@@ -81,6 +81,47 @@ fun File.generateInitialCommitAddVersionTagAndAddNewCommit(
                 )))
 }
 
+// `this` is `projectDir`
+fun File.generateInitialCommitAddVersionTag(
+    doBefore: Git.() -> Unit = {},
+    doAfter: Git.() -> Unit = {},
+): List<Commit> {
+    createGitIgnore()
+    val git: Git = Git.init().setDirectory(this).call()
+    doBefore(git)
+    resolve("initial.txt").createNewFile()
+    git.add().addFilepattern(".").call()
+    val commit = git.commit().setMessage("Initial commit").call()
+    git.tag().setObjectId(commit).setName(resolve("last-tag.txt").readLines().first()).call()
+    resolve("library/last-tag.txt").apply {
+        if (exists()) git.tag().setObjectId(commit).setName(readLines().first()).call()
+    }
+    doAfter(git)
+
+    return listOf(
+        Commit(
+            message = "Add new2",
+            fullMessage = "Add new2",
+            hash = "f099fed42808c387d38f8e34934d4ef2d6f3c2c5",
+            timestampEpochSecond = Instant.now().epochSecond,
+            tags = emptyList()),
+        Commit(
+            message = "Add new",
+            fullMessage = "Add new",
+            hash = "0d9be4d12a836435c91fc38e0d64bcd3c14c89c8",
+            timestampEpochSecond = Instant.now().epochSecond,
+            tags = emptyList()),
+        Commit(
+            message = "Initial commit",
+            fullMessage = "Initial commit",
+            hash = "e2592aafc259bd797e32f1ea4fe60cc7c0698f70",
+            timestampEpochSecond = Instant.now().epochSecond,
+            tags =
+                listOf(
+                    Tag(name = "1.0.0", refName = "refs/tags/1.0.0"),
+                )))
+}
+
 fun File.createGitIgnore() {
     resolve(".gitignore").apply {
         createNewFile()